stock-analyst/requirements.txt
pydantic>=2.7.0
httpx>=0.27.0
typer>=0.12.0
rich>=13.7.0
pandas>=2.2.0
python-dotenv>=1.0.1
